del file
